Create `span` instead of `rb` elements
Closes #237. 

Empirically, it appears that browsers (Firefox and Chrome) do not implement layout on the currently deprecated `rb` element in the same way as on other elements. Specifically, they seem not to extend the padding rectangle to include the padded rectangles of child `span`s. Replacing the `rb` elements with `span` elements works around this. 

If at some point `rb` is un-deprecated, and this feature/bug of layout is improved, this change should be reverted.